+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnChanges, SimpleChanges } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 
 @Component({
@@ -8,16 +8,4 @@ import { NavigationEnd, Router } from '@angular/router';
 })
 export class AppComponent {
   title = 'dart-scorer';
-  showGameComponents: boolean = false;
-
-  constructor(private router: Router) { }
-
-  ngOnInit(): void {
-    this.router.events.subscribe(event => {
-      if (event instanceof NavigationEnd) {
-        const currentURL = event.url;
-        this.showGameComponents = currentURL === '/singleOut' || currentURL === '/doubleOut' || currentURL === '/aroundTheClock';
-      }
-    });
-  }
 }